+++ Launcher/Utility/AutoShadersGen3.cs
@@ -35,16 +35,18 @@ internal class AutoShadersGen3
         // There must be a better way of doing this that doesn't involve try-catch, but I'm tired...
         try
         {
-            if (Directory.GetFiles(destinationShadersFolder) == Array.Empty<string>())
-            {
-                Debug.WriteLine("Folder exists but contains no shaders, proceeding");
-                shaderGen(files, full_jms_path, counter, destinationShadersFolder, BaseDirectory, gameType);
-            }
-            else
+            if (!(Directory.GetFiles(destinationShadersFolder) == Array.Empty<string>()))
             {
                 Debug.WriteLine("Shaders already exist!");
-                MessageBox.Show("Shaders for this model already exist, skipping shader generation!", "Shader Gen. Warning", MessageBoxButtons.OK, MessageBoxIcon.Information);
-            }
+                if (MessageBox.Show("Shaders for this model already exist!\nWould you like to generate any missing shaders?", "Shader Gen. Warning", MessageBoxButtons.YesNo, MessageBoxIcon.Information) == DialogResult.Yes)
+                {
+                    shaderGen(files, full_jms_path, counter, destinationShadersFolder, BaseDirectory, gameType);
+                }
+                else
+                {
+                    return true;
+                }
+            }   
         }
         catch (DirectoryNotFoundException)
         {
@@ -171,33 +173,36 @@ internal class AutoShadersGen3
             foreach (string shader in shaders)
             {
                 string shaderName = shader + ".shader";
-                try
+                if(!File.Exists(Path.Combine(destinationShadersFolder, shaderName)))
                 {
-                    File.Copy(defaultShaderLocation, Path.Combine(destinationShadersFolder, shaderName));
-                }
-                catch (FileNotFoundException)
-                {
-                    // Will probably only occur if user somehow deletes default.shader after the check for its existence occurs,
-                    // but before shaders are generated
-                    if (MessageBox.Show("Unable to find shader to copy from!\nThis really shouldn't have happened.\nPress OK to try again, or Cancel to skip shader generation.", "Shader Gen. Error", MessageBoxButtons.OKCancel, MessageBoxIcon.Warning) == DialogResult.OK)
+                    try
                     {
-                        if (gameType == "H3")
+                        File.Copy(defaultShaderLocation, Path.Combine(destinationShadersFolder, shaderName));
+                    }
+                    catch (FileNotFoundException)
+                    {
+                        // Will probably only occur if user somehow deletes default.shader after the check for its existence occurs,
+                        // but before shaders are generated
+                        if (MessageBox.Show("Unable to find shader to copy from!\nThis really shouldn't have happened.\nPress OK to try again, or Cancel to skip shader generation.", "Shader Gen. Error", MessageBoxButtons.OKCancel, MessageBoxIcon.Warning) == DialogResult.OK)
                         {
-                            File.WriteAllBytes(defaultShaderLocation, ToolkitLauncher.Utility.Resources.defaultH3);
+                            if (gameType == "H3")
+                            {
+                                File.WriteAllBytes(defaultShaderLocation, ToolkitLauncher.Utility.Resources.defaultH3);
+                            }
+                            else
+                            {
+                                File.WriteAllBytes(defaultShaderLocation, ToolkitLauncher.Utility.Resources.defaultODST);
+                            }
+                            counter = 0;
+                            shaderGen(files, full_jms_path, counter, destinationShadersFolder, BaseDirectory, gameType);
+                            break;
                         }
                         else
                         {
-                            File.WriteAllBytes(defaultShaderLocation, ToolkitLauncher.Utility.Resources.defaultODST);
+                            break;
                         }
-                        counter = 0;
-                        shaderGen(files, full_jms_path, counter, destinationShadersFolder, BaseDirectory, gameType);
-                        break;
                     }
-                    else
-                    {
-                        break;
-                    }
-                }
+                }       
             }
         }
         return true;
